Refactored all classes
