+++ Moblima/MovieBookerInterface.java
@@ -1,0 +1,25 @@
+package Moblima;
+
+import java.io.FileNotFoundException;
+
+public interface MovieBookerInterface {
+
+	void showAllMovies();
+
+	void searchMovie();
+
+	void showShowTimes();
+
+	void showAllMoviesTicket();
+
+	void createRatingReview();
+
+	void bookingMenu();
+
+	void showBookingHist();
+
+	void addExamples();
+
+
+	
+}
